compiler: Lower buffering from equations to clusters level
