+++ devito/passes/clusters/__init__.py
@@ -1,4 +1,5 @@
 from .utils import *  # noqa
+from .buffering import *  # noqa
 from .aliases import *  # noqa
 from .cse import *  # noqa
 from .factorization import *  # noqa
