+++ devito/passes/equations/__init__.py
@@ -1,2 +1,1 @@
 from .linearity import *  # noqa
-from .buffering import *  # noqa
